feat: address review comments and move the code to service file
