+++ redfish/internal/controller/http/v1/handler/routes.go
@@ -2,12 +2,9 @@
 package v1
 
 import (
-	"encoding/xml"
 	"errors"
 	"fmt"
 	"net/http"
-	"os"
-	"sync"
 	"time"
 
 	"github.com/gin-gonic/gin"
@@ -19,66 +16,6 @@ import (
 	"github.com/device-management-toolkit/console/redfish/internal/usecase"
 )
 
-var (
-	metadataXML    string
-	metadataLoaded bool
-	metadataMutex  sync.Mutex
-)
-
-// loadMetadata loads metadata.xml from the generated folder with XML validation.
-// This function is thread-safe and ensures metadata is loaded only once.
-func loadMetadata() {
-	metadataMutex.Lock()
-
-	// Double-check locking pattern
-	if metadataLoaded {
-		metadataMutex.Unlock()
-
-		return
-	}
-
-	data, err := os.ReadFile(metadataFilePath)
-	if err != nil {
-		log.Warnf("Could not load metadata.xml from %s: %v", metadataFilePath, err)
-
-		metadataXML = ""
-		metadataLoaded = true
-
-		metadataMutex.Unlock()
-
-		return
-	}
-
-	metadataXML = string(data)
-
-	// Validate XML
-	if err := validateMetadataXML(metadataXML); err != nil {
-		metadataMutex.Unlock()
-		log.Fatalf("Invalid metadata.xml at %s: %v", metadataFilePath, err)
-	}
-
-	log.Infof("metadata.xml loaded from %s and validation passed", metadataFilePath)
-
-	metadataLoaded = true
-
-	metadataMutex.Unlock()
-}
-
-// validateMetadataXML checks if the provided XML string is well-formed.
-func validateMetadataXML(xmlData string) error {
-	if xmlData == "" {
-		return nil // Empty is acceptable
-	}
-
-	var doc interface{}
-
-	if err := xml.Unmarshal([]byte(xmlData), &doc); err != nil {
-		return fmt.Errorf("XML parsing failed: %w", err)
-	}
-
-	return nil
-}
-
 const (
 	// Task state constants from Redfish Task.v1_8_0 specification
 	taskStateCompleted = "Completed"
@@ -87,11 +24,6 @@ const (
 	msgIDBaseSuccess      = "Base.1.22.0.Success"
 	msgIDBaseGeneralError = "Base.1.22.0.GeneralError"
 )
-
-// metadataFilePath is the location of the metadata.xml file
-// The metadata.xml is generated by: make rf-metadata-generate
-// Source: redfish/openapi/infra/metadata-generate.py
-const metadataFilePath = "redfish/internal/controller/http/v1/generated/metadata.xml"
 
 // RedfishServer implements the Redfish API handlers
 // Add dependencies here if needed (e.g., usecase, presenter, etc.)
@@ -122,41 +54,6 @@ func Int64Ptr(i int64) *int64 {
 // SystemTypePtr creates a pointer to a ComputerSystemSystemType value.
 func SystemTypePtr(st generated.ComputerSystemSystemType) *generated.ComputerSystemSystemType {
 	return &st
-}
-
-// GetRedfishV1 returns the service root
-func (s *RedfishServer) GetRedfishV1(c *gin.Context) {
-	serviceRoot := generated.ServiceRootServiceRoot{
-		OdataContext:   StringPtr("/redfish/v1/$metadata#ServiceRoot.ServiceRoot"),
-		OdataId:        StringPtr("/redfish/v1"),
-		OdataType:      StringPtr("#ServiceRoot.v1_19_0.ServiceRoot"),
-		Id:             "RootService",
-		Name:           "Root Service",
-		RedfishVersion: StringPtr("1.19.0"),
-		Systems: &generated.OdataV4IdRef{
-			OdataId: StringPtr("/redfish/v1/Systems"),
-		},
-	}
-	c.JSON(http.StatusOK, serviceRoot)
-}
-
-// GetRedfishV1Metadata returns the OData CSDL metadata document describing the service's data model.
-// Per Redfish specification, this endpoint is accessible without authentication.
-func (s *RedfishServer) GetRedfishV1Metadata(c *gin.Context) {
-	// Ensure metadata is loaded
-	loadMetadata()
-
-	// Read metadata safely
-	metadataMutex.Lock()
-
-	metadata := metadataXML
-
-	metadataMutex.Unlock()
-
-	// Set Redfish-compliant headers
-	c.Header(headerContentType, contentTypeXML)
-	c.Header(headerODataVersion, odataVersion)
-	c.String(http.StatusOK, metadata)
 }
 
 // GetRedfishV1Systems returns the computer systems collection
